Move user address normalization into separate module
With upcoming changes we will introduce support for normalizing kernel
addresses. Given that it's mostly independent of the existing
normalization functionality, let's introduce a bit of a split between
logic concerned with user address normalization and that for kernel
address normalization.
Start by moving most bits pertaining user address normalization into a
new sub-module, user.
